Remove unnecesary namespaces
