+++ app/lib/crypto/mnemonic.ts
@@ -1,10 +1,18 @@
 const bip39 = require("bip39")
 
-export namespace Mnemonic {
-  export const validate = (mnemonics: string): boolean => {
-    return !bip39.validateMnemonic(mnemonics)
-  }
-  export const generate = () => {
-    return bip39.generateMnemonic()
-  }
+/**
+ * Validate mnemonic
+ * @param {string} mnemonics
+ * @returns {boolean}
+ */
+export const validate = (mnemonics: string): boolean => {
+  return !bip39.validateMnemonic(mnemonics)
+}
+
+/**
+ * Generate mnemonic
+ * @returns {string}
+ */
+export const generate = () => {
+  return bip39.generateMnemonic()
 }
