+++ app/lib/crypto/publicKey.ts
@@ -1,7 +1,7 @@
 const secp256k1 = require("secp256k1")
 
-export namespace PublicKey {
-  export const create = (privateKey: Buffer, compressed = true): Buffer => {
-    return secp256k1.publicKeyCreate(privateKey, compressed)
-  }
+export type PublicKey = Buffer
+
+export const create = (privateKey: Buffer, compressed = true): PublicKey => {
+  return secp256k1.publicKeyCreate(privateKey, compressed)
 }
